+++ NEWS.md
@@ -1,4 +1,12 @@
+## grumpy 0.1.1 _2021-09-16_
+
+    * Add some initial news file checks
+        * check NEWS.md exists
+        * check NEWS.md version matches toml version
+        * check NEWS.md has correct format
+
 ## grumpy 0.1.0 _2021-09-16_
+
     * Intial commit
     * Add skeleton functionality for checks with reporting
     * Add skeleton functionality for register plugin style checks